# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample online-shop server",
        "title": "Online shop API Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Возвращает список всех товаров",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получить все товары",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/types.Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Добавляет новый товар в список",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Создать новый товар",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Информация о товаре",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ProductData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Возвращает информацию о товаре по его ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получить товар по ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Product"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Обновляет данные товара по ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Обновить существующий товар",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные товара",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.Product"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Удаляет товар по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Удалить товар",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.ProductData"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}/category": {
            "post": {
                "description": "Добавляет новую категорию для существующего товара",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Добавить категорию товара",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название категории",
                        "name": "category",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/products/{id}/cost": {
            "post": {
                "description": "Добавляет новую стоимость для существующего товара",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Добавить стоимость товара",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Токен",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Информация о стоимости",
                        "name": "cost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.ProductCost"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/types.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.Category": {
            "type": "object",
            "properties": {
                "categoryID": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "description": {
                    "description": "Description of the category",
                    "type": "string"
                },
                "name": {
                    "description": "Category name",
                    "type": "string"
                }
            }
        },
        "types.Characteristic": {
            "type": "object",
            "properties": {
                "category": {
                    "description": "Name of category",
                    "type": "string"
                },
                "characteristicID": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "name": {
                    "description": "Name of the characteristic",
                    "type": "string"
                }
            }
        },
        "types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "types.Manufacturer": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "Description of the manufacturer",
                    "type": "string"
                },
                "manufacturerID": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "name": {
                    "description": "Manufacturer name",
                    "type": "string"
                }
            }
        },
...
        "types.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProductCategory"
                    }
                },
                "characteristic": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProductCharacteristic"
                    }
                },
                "cost": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/types.ProductCost"
                    }
                },
                "description": {
                    "description": "Product description",
                    "type": "string"
                },
                "imageLink": {
                    "description": "URL to the product image",
                    "type": "string"
                },
                "manufacturer": {
                    "description": "Associations",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Manufacturer"
                        }
                    ]
                },
                "manufacturer_id": {
                    "description": "Foreign key to Manufacturer",
                    "type": "integer"
                },
                "name": {
                    "description": "Product name",
                    "type": "string"
                },
                "product_id": {
                    "description": "Primary key",
                    "type": "integer"
                },
                "status": {
                    "description": "Status",
                    "type": "string"
                }
            }
        },
        "types.ProductCategory": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/types.Category"
                }
            }
        },
        "types.ProductCharacteristic": {
            "type": "object",
            "properties": {
                "characteristic": {
                    "$ref": "#/definitions/types.Characteristic"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "types.ProductCost": {
            "type": "object",
            "properties": {
                "costId": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "productId": {
                    "type": "integer"
                },
                "startTimeStamp": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "types.ProductData": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "imageLink": {
                    "type": "string"
                },
                "manufacturer_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "types.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
...
    },
...
}